[DIDA] Presentation 레이어 수정중 중간 머지

+++ feature/hot-user/src/main/java/com/dida/hot_user/HotUserBindingAdapters.kt
@@ -6,7 +6,7 @@ import androidx.annotation.RequiresApi
 import androidx.databinding.BindingAdapter
 
 @BindingAdapter("cardCount")
-fun TextView.bindCardCount(count: Long) {
+fun TextView.bindCardCount(count: Int) {
     val view = this
     view.text = "$count 작품"
 }
